refactor: move imports page under pages
