+++ frontend/tests/imports/importPreview.test.ts
@@ -6,9 +6,9 @@ import type { AccountsOverview } from '@/api/accounts'
 import type { Category } from '@/api/categories'
 import type { Currency } from '@/api/currency'
 import type { Institution } from '@/api/institutions'
-import { CREATE_ACCOUNT_VALUE, CREATE_CATEGORY_VALUE, EMPTY_COLUMN_MAP } from '@/imports/constants'
-import type { ImportFileDraft } from '@/imports/types'
-import { buildImportPreviewRows } from '@/imports/utils'
+import { CREATE_ACCOUNT_VALUE, CREATE_CATEGORY_VALUE, EMPTY_COLUMN_MAP } from '@/pages/imports/constants'
+import type { ImportFileDraft } from '@/pages/imports/types'
+import { buildImportPreviewRows } from '@/pages/imports/utils'
 
 const currencies: Currency[] = [
   { id: 'USD', name: 'US Dollar', symbol: '$', minor_unit_exponent: 2 },
